requirements.txt
python-dateutil~=2.9.0
sqlalchemy~=2.0.36
scrapy~=2.12.0
matplotlib~=3.8.4
jieba~=0.42.1
gensim~=4.3.3
beautifulsoup4~=4.12.3
itemadapter~=0.10.0
PyQt5~=5.15.4
pandas~=2.2.3
numpy~=1.26.4
matplotlib~=3.8.4
seaborn~=0.13.2
wordcloud~=1.9.3
flask~=3.0.3
redis-py-cluster~=2.1.3
redis~=3.5.3
apscheduler~=3.11.0
mmh3~=5.0.1
bitarray~=3.0.0